+++ dddplus-visualization/src/main/java/io/github/dddplus/ast/report/CallGraphReport.java
@@ -5,10 +5,7 @@
  */
 package io.github.dddplus.ast.report;
 
-import io.github.dddplus.ast.model.PackageCrossRefEntry;
-import io.github.dddplus.ast.model.ReverseEngineeringModel;
 import io.github.dddplus.ast.model.CallGraphEntry;
-import io.github.dddplus.ast.model.KeyModelEntry;
 import lombok.Data;
 
 import java.util.*;
@@ -18,97 +15,54 @@ import java.util.*;
  */
 @Data
 public class CallGraphReport {
-    private final ReverseEngineeringModel model;
     private List<CallGraphEntry> entries = new ArrayList<>();
-    private Set<PackageCrossRefEntry> packageCrossRefEntries = new TreeSet<>();
-    private int calls = 0;
-    private int parsedCalls = 0;
-
-    public CallGraphReport(ReverseEngineeringModel model) {
-        this.model = model;
-    }
-
-    public boolean interestedInMethod(String declarationClazz, String methodName) {
-        return model.getKeyModelReport().hasKeyMethod(declarationClazz, methodName);
-    }
-
-    public double callCoverage() {
-        if (calls == 0) {
-            // user might disable call graph
-            return 0;
-        }
-
-        return parsedCalls * 100 / calls;
-    }
-
-    public void incrementCalls() {
-        calls++;
-    }
-
-    private void incrementParsedCalls() {
-        parsedCalls++;
-    }
 
     public List<CallGraphEntry> sortedEntries() {
         Collections.sort(entries, Comparator.comparing(c -> c.getCallerClazz() + c.getCallerMethod()));
         return entries;
     }
 
-    public boolean isKeyModel(String clazz) {
-        return model.getKeyModelReport().containsActor(clazz);
-    }
-
-    private Set<String> calleeClasses() {
-        Set<String> r = new TreeSet<>();
-        for (CallGraphEntry entry : entries) {
-            r.add(entry.getCalleeClazz());
-        }
-        return r;
-    }
-
     public Collection<Record> calleeRecords() {
         List<Record> records = new ArrayList<>();
-        for (String calleeClass : calleeClasses()) {
+        Set<String> calleeClasses = new TreeSet<>();
+        for (CallGraphEntry entry : entries) {
+            calleeClasses.add(entry.getCalleeClazz());
+        }
+        for (String calleeClass : calleeClasses) {
             Record record = new Record(calleeClass);
-            record.addMethods(model.getKeyModelReport().keyModelEntryOfActor(calleeClass).realKeyMethods());
-
+            for (CallGraphEntry entry : entries) {
+                if (!entry.getCalleeClazz().equals(calleeClass)) {
+                    continue;
+                }
+                record.getMethods().add(entry.getCalleeMethod());
+            }
             records.add(record);
         }
         return records;
     }
 
-    public void register(String callerClazz, String callerMethod, String calleeClazz, String calleeMethod) {
-        incrementParsedCalls();
-
-        if (!model.getKeyModelReport().containsActor(calleeClazz)) {
-            // 被调用的方法不是我们标注的，例如 BigDecimal::add
-            return;
+    public Collection<Record> callerRecords() {
+        List<Record> records = new ArrayList<>();
+        Set<String> callerClasses = new TreeSet<>();
+        for (CallGraphEntry entry : entries) {
+            callerClasses.add(entry.getCallerClazz());
         }
+        for (String callerClazz : callerClasses) {
+            Record record = new Record(callerClazz);
+            for (CallGraphEntry entry : entries) {
+                if (!entry.getCallerClazz().equals(callerClazz)) {
+                    continue;
+                }
 
-        // 被动注册：callee model method
-        KeyModelEntry calleeModel = model.getKeyModelReport().keyModelEntryOfActor(calleeClazz);
-        if (!calleeModel.hasKeyMethod(calleeMethod)) {
-            calleeModel.registerMethodFodCallGraph(calleeMethod);
+                record.addMethod(entry.getCallerMethod());
+            }
+            records.add(record);
         }
-        // 被动注册：caller model method
-        KeyModelEntry callerModel = model.getKeyModelReport().keyModelEntryOfActor(callerClazz);
-        if (callerModel != null && !callerModel.hasKeyMethod(callerMethod)) {
-            callerModel.registerMethodFodCallGraph(callerMethod);
-        }
-
-        entries.add(new CallGraphEntry(callerClazz, callerMethod, calleeClazz, calleeMethod));
+        return records;
     }
 
-    public void addPackageCrossRef(String callerPackage, String calleePackage) {
-        if (callerPackage.equals(calleePackage)) {
-            // only add cross packages relations
-            return;
-        }
-        if (!model.hasPackage(calleePackage) || !model.hasPackage(callerPackage)) {
-            return;
-        }
-
-        packageCrossRefEntries.add(new PackageCrossRefEntry(callerPackage, calleePackage));
+    public void register(CallGraphEntry entry) {
+        entries.add(entry);
     }
 
     @Data
@@ -120,8 +74,12 @@ public class CallGraphReport {
             this.clazz = clazz;
         }
 
-        void addMethods(Set<String> methods) {
-            this.methods.addAll(methods);
+        public String dotNode() {
+            return clazz.replaceAll("\\.", "_");
+        }
+
+        void addMethod(String method) {
+            this.methods.add(method);
         }
     }
 }