+++ dddplus-visualization/src/main/java/io/github/dddplus/ast/report/KeyModelReport.java
@@ -52,16 +52,6 @@ public class KeyModelReport {
         }
     }
 
-    public boolean hasKeyMethod(String declarationClazz, String methodName) {
-        for (KeyModelEntry modelEntry : data.values()) {
-            if (modelEntry.getClassName().equals(declarationClazz) && modelEntry.hasKeyMethod(methodName)) {
-                return true;
-            }
-        }
-
-        return false;
-    }
-
     public KeyModelEntry getOrCreateKeyModelEntryForActor(String actor) {
         if (!data.containsKey(actor)) {
             data.put(actor, new KeyModelEntry(actor));